test(order): 주문 테스트 패키지 구조 개선
