+++ src/test/java/com/musinsa/orders/domain/order/MoneyTest.java
@@ -1,4 +1,4 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.order;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
