+++ src/test/java/com/musinsa/orders/domain/order/OrderTest.java
@@ -1,11 +1,13 @@
-package com.musinsa.orders.domain;
+package com.musinsa.orders.domain.order;
 
 import static com.musinsa.orders.Fixtures.createOrder;
 import static com.musinsa.orders.Fixtures.createOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.tuple;
 
 import com.musinsa.orders.domain.order.Money;
 import com.musinsa.orders.domain.order.Order;
+import com.musinsa.orders.domain.order.ProductName;
 import com.musinsa.orders.domain.order.ShippingFeePolicy;
 import com.musinsa.orders.infra.order.AmountShippingFeePolicy;
 import java.util.List;
@@ -13,7 +15,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ShippingFeePolicyTest {
+class OrderTest {
 
   private ShippingFeePolicy shippingFeePolicy;
 
@@ -22,29 +24,24 @@ class ShippingFeePolicyTest {
     shippingFeePolicy = new AmountShippingFeePolicy();
   }
 
-  @DisplayName("5만원 미만 주문금액에 대해 유료 배송비 2500원을 반환한다")
   @Test
-  void calculateShippingFee() {
-    Order order = createOrder(1L, List.of(
+  @DisplayName("주문생성")
+  void placeOrder() {
+    Order actual = createOrder(1L, List.of(
         createOrderLineItem(1L, "신발A", 15_000L),
         createOrderLineItem(2L, "신발B", 16_000L),
         createOrderLineItem(3L, "신발C", 17_000L)
-    ));
+    ), shippingFeePolicy);
 
-    Money actual = shippingFeePolicy.calculateShippingFee(order);
-    assertThat(actual).isEqualTo(Money.from(2_500L));
+    assertThat(actual).isNotNull();
+    assertThat(actual.orderLineItems()).hasSize(3);
+    assertThat(actual.shippingFee()).isEqualTo(Money.from(2_500L));
+    assertThat(actual.orderLineItems()).extracting("productId", "name", "price")
+        .contains(
+            tuple(1L, ProductName.from("신발A"), Money.from(15_000L)),
+            tuple(2L, ProductName.from("신발B"), Money.from(16_000L)),
+            tuple(3L, ProductName.from("신발C"), Money.from(17_000L))
+        );
   }
 
-  @DisplayName("5만원 이상 주문금액에 대해 배송비 0원(무료배송금액)을 반환한다")
-  @Test
-  void calculateShippingFee_Free() {
-    Order order = createOrder(1L, List.of(
-        createOrderLineItem(1L, "셔츠A", 40_000L),
-        createOrderLineItem(2L, "셔츠B", 50_000L),
-        createOrderLineItem(3L, "셔츠C", 60_000L)
-    ));
-
-    Money actual = shippingFeePolicy.calculateShippingFee(order);
-    assertThat(actual).isEqualTo(Money.ZERO);
-  }
 }
